feat: create separate packages
