+++ packages/contracts/hardhat.config.ts
@@ -1,0 +1,38 @@
+import * as dotenv from 'dotenv';
+
+import { HardhatUserConfig, task } from 'hardhat/config';
+import '@nomiclabs/hardhat-waffle';
+import '@typechain/hardhat';
+import 'hardhat-gas-reporter';
+import 'solidity-coverage';
+
+dotenv.config();
+
+task('accounts', 'Prints the list of accounts', async (taskArgs, hre) => {
+  const accounts = await hre.ethers.getSigners();
+  console.log(taskArgs);
+  for (const account of accounts) {
+    console.log(account.address);
+  }
+});
+
+const config: HardhatUserConfig = {
+  solidity: {
+    version: '0.8.4',
+    settings: {
+      optimizer: {
+        enabled: true,
+        runs: 200,
+      },
+    },
+  },
+
+  networks: {
+    mumbai: {
+      url: process.env.MUMBAI_URL || '',
+      accounts: process.env.PRIVATE_KEY !== undefined ? [process.env.PRIVATE_KEY] : [],
+    },
+  },
+};
+
+export default config;
