+++ packages/contracts/scripts/test/index.ts
